+++ src/__tests__/container.js
@@ -27,6 +27,78 @@ test('container#eachInside', (t) => {
     t.equal(str, 'h1h2h3h4');
 });
 
+test('container#eachAttribute', (t) => {
+    t.plan(1);
+    let out = parse('[href][class].class', (selectors) => {
+        selectors.eachAttribute((attr) => {
+            if (attr.attribute === 'class') {
+                attr.removeSelf();
+            }
+        });
+    });
+    t.equal(out, '[href].class');
+});
+
+test('container#eachClass', (t) => {
+    t.plan(1);
+    let out = parse('.one, .two, .three:not(.four, .five)', (selectors) => {
+        selectors.eachClass((className) => {
+            className.value = className.value.slice(0, 1);
+        });
+    });
+    t.equal(out, '.o, .t, .t:not(.f, .f)');
+});
+
+test('container#eachComment', (t) => {
+    t.plan(1);
+    let out = parse('.one/*test*/.two', (selectors) => {
+        selectors.eachComment((comment) => {
+            comment.removeSelf();
+        });
+    });
+    t.equal(out, '.one.two');
+});
+
+test('container#eachId', (t) => {
+    t.plan(1);
+    let out = parse('h1#one, h2#two', (selectors) => {
+        selectors.eachId((id) => {
+            id.value = id.value.slice(0, 1);
+        });
+    });
+    t.equal(out, 'h1#o, h2#t');
+});
+
+test('container#eachPseudo', (t) => {
+    t.plan(1);
+    let out = parse('a:before, a:after', (selectors) => {
+        selectors.eachPseudo((pseudo) => {
+            pseudo.value = pseudo.value.slice(0, 2);
+        });
+    });
+    t.equal(out, 'a:b, a:a');
+});
+
+test('container#eachTag', (t) => {
+    t.plan(1);
+    let out = parse('1 2 3', (selectors) => {
+        selectors.eachTag((tag) => {
+            tag.value = 'h' + tag.value;
+        });
+    });
+    t.equal(out, 'h1 h2 h3');
+});
+
+test('container#eachUniversal', (t) => {
+    t.plan(1);
+    let out = parse('*.class,*.class,*.class', (selectors) => {
+        selectors.eachUniversal((universal) => {
+            universal.removeSelf();
+        });
+    });
+    t.equal(out, '.class,.class,.class');
+});
+
 test('container#map', (t) => {
     t.plan(1);
     parse('1 2 3', (selectors) => {